Create 'getDocsURL' helper function
