--- src/hooks/useDocsURL.ts
@@ -1,41 +1,0 @@
-import { useState, useEffect } from 'react';
-
-import { frameworkMap } from '../common/constants';
-import { Strings } from '../resources/Strings';
-import { Framework } from '../types/types';
-
-type SupportedPerfdocsFramework = 'talos' | 'awsy' | 'devtools';
-const supportedPerfdocsFrameworks: Record<SupportedPerfdocsFramework, string> =
-  {
-    talos: 'talos',
-    awsy: 'awsy',
-    devtools: 'performance-tests-overview',
-  };
-
-const useDocsURL = (suite: string, framework_id: Framework['id']) => {
-  const [docsURL, setdocsURL] = useState('');
-  const [isLinkSupported, setIsLinkSupported] = useState(true);
-
-  useEffect(() => {
-    const framework = frameworkMap[framework_id];
-    const supportedFramework =
-      supportedPerfdocsFrameworks[framework as SupportedPerfdocsFramework];
-    const urlReadySuite = suite.replace(/:|\s|\.|_/g, '-').toLowerCase();
-
-    if (framework_id === 12 && supportedFramework) {
-      setdocsURL(
-        `${Strings.urls.docs.baseURL}/devtools/tests/${supportedFramework}.html#${urlReadySuite}`,
-      );
-    } else if (supportedFramework) {
-      setdocsURL(
-        `${Strings.urls.docs.baseURL}/testing/perfdocs/${supportedFramework}.html#${urlReadySuite}`,
-      );
-    } else {
-      setIsLinkSupported(false);
-    }
-  }, [suite, framework_id]);
-
-  return { docsURL, isLinkSupported };
-};
-
-export default useDocsURL;
